+++ Ladybird/Utilities.h
@@ -1,0 +1,13 @@
+/*
+ * Copyright (c) 2022, Andreas Kling <kling@serenityos.org>
+ *
+ * SPDX-License-Identifier: BSD-2-Clause
+ */
+
+#pragma once
+
+#include <AK/String.h>
+#include <QString>
+
+AK::String akstring_from_qstring(QString const&);
+QString qstring_from_akstring(AK::String const&);
