+++ Ladybird/WebView.cpp
@@ -13,7 +13,9 @@
 #include "EventLoopPluginQt.h"
 #include "FontPluginQt.h"
 #include "ImageCodecPluginLadybird.h"
+#include "PageClientLadybird.h"
 #include "RequestManagerQt.h"
+#include "Utilities.h"
 #include <AK/Assertions.h>
 #include <AK/ByteBuffer.h>
 #include <AK/Format.h>
@@ -67,300 +69,13 @@
 #include <QToolTip>
 #include <QVBoxLayout>
 
-AK::String akstring_from_qstring(QString const& qstring)
-{
-    return AK::String(qstring.toUtf8().data());
-}
-
-QString qstring_from_akstring(AK::String const& akstring)
-{
-    return QString::fromUtf8(akstring.characters(), akstring.length());
-}
-
 String s_serenity_resource_root;
-
-class HeadlessBrowserPageClient final : public Web::PageClient {
-public:
-    static NonnullOwnPtr<HeadlessBrowserPageClient> create(WebView& view)
-    {
-        return adopt_own(*new HeadlessBrowserPageClient(view));
-    }
-
-    Web::Page& page() { return *m_page; }
-    Web::Page const& page() const { return *m_page; }
-
-    Web::Layout::InitialContainingBlock* layout_root()
-    {
-        auto* document = page().top_level_browsing_context().active_document();
-        if (!document)
-            return nullptr;
-        return document->layout_node();
-    }
-
-    void load(AK::URL const& url)
-    {
-        if (!url.is_valid())
-            return;
-
-        page().load(url);
-    }
-
-    void paint(Gfx::IntRect const& content_rect, Gfx::Bitmap& target)
-    {
-        Gfx::Painter painter(target);
-
-        if (auto* document = page().top_level_browsing_context().active_document())
-            document->update_layout();
-
-        painter.fill_rect({ {}, content_rect.size() }, palette().base());
-
-        auto* layout_root = this->layout_root();
-        if (!layout_root) {
-            return;
-        }
-
-        Web::PaintContext context(painter, palette(), content_rect.top_left());
-        context.set_should_show_line_box_borders(m_should_show_line_box_borders);
-        context.set_viewport_rect(content_rect);
-        context.set_has_focus(true);
-        layout_root->paint_all_phases(context);
-    }
-
-    void setup_palette(Core::AnonymousBuffer theme_buffer)
-    {
-        m_palette_impl = Gfx::PaletteImpl::create_with_anonymous_buffer(theme_buffer);
-    }
-
-    void set_viewport_rect(Gfx::IntRect rect)
-    {
-        m_viewport_rect = rect;
-        page().top_level_browsing_context().set_viewport_rect(rect);
-    }
-
-    // ^Web::PageClient
-    virtual Gfx::Palette palette() const override
-    {
-        return Gfx::Palette(*m_palette_impl);
-    }
-
-    virtual Gfx::IntRect screen_rect() const override
-    {
-        // FIXME: Return the actual screen rect.
-        return m_viewport_rect;
-    }
-
-    Gfx::IntRect viewport_rect() const
-    {
-        return m_viewport_rect;
-    }
-
-    virtual Web::CSS::PreferredColorScheme preferred_color_scheme() const override
-    {
-        return m_preferred_color_scheme;
-    }
-
-    virtual void page_did_change_title(String const& title) override
-    {
-        emit m_view.title_changed(title.characters());
-    }
-
-    virtual void page_did_start_loading(AK::URL const& url) override
-    {
-        emit m_view.load_started(url);
-    }
-
-    virtual void page_did_finish_loading(AK::URL const&) override
-    {
-        initialize_js_console();
-        m_console_client->send_messages(0);
-    }
-
-    void initialize_js_console()
-    {
-        auto* document = page().top_level_browsing_context().active_document();
-        auto realm = document->realm().make_weak_ptr();
-        if (m_realm && m_realm.ptr() == realm.ptr())
-            return;
-
-        m_realm = realm;
-
-        auto& console_object = *document->realm().intrinsics().console_object();
-        m_console_client = make<Ladybird::ConsoleClient>(console_object.console(), *realm, m_view);
-        console_object.console().set_client(*m_console_client.ptr());
-    }
-
-    virtual void page_did_change_selection() override
-    {
-    }
-
-    virtual void page_did_request_cursor_change(Gfx::StandardCursor cursor) override
-    {
-        switch (cursor) {
-        case Gfx::StandardCursor::Hand:
-            m_view.setCursor(Qt::PointingHandCursor);
-            break;
-        case Gfx::StandardCursor::IBeam:
-            m_view.setCursor(Qt::IBeamCursor);
-            break;
-        case Gfx::StandardCursor::Arrow:
-        default:
-            m_view.setCursor(Qt::ArrowCursor);
-            break;
-        }
-    }
-
-    virtual void page_did_request_context_menu(Gfx::IntPoint const&) override
-    {
-    }
-
-    virtual void page_did_request_link_context_menu(Gfx::IntPoint const&, AK::URL const&, String const&, unsigned) override
-    {
-    }
-
-    virtual void page_did_request_image_context_menu(Gfx::IntPoint const&, AK::URL const&, String const&, unsigned, Gfx::Bitmap const*) override
-    {
-    }
-
-    virtual void page_did_click_link(AK::URL const&, String const&, unsigned) override
-    {
-    }
-
-    virtual void page_did_middle_click_link(AK::URL const&, String const&, unsigned) override
-    {
-    }
-
-    virtual void page_did_enter_tooltip_area(Gfx::IntPoint const& content_position, String const& tooltip) override
-    {
-        auto widget_position = m_view.to_widget(content_position);
-        QToolTip::showText(
-            m_view.mapToGlobal(QPoint(widget_position.x(), widget_position.y())),
-            qstring_from_akstring(tooltip),
-            &m_view);
-    }
-
-    virtual void page_did_leave_tooltip_area() override
-    {
-        QToolTip::hideText();
-    }
-
-    virtual void page_did_hover_link(AK::URL const& url) override
-    {
-        emit m_view.link_hovered(url.to_string().characters());
-    }
-
-    virtual void page_did_unhover_link() override
-    {
-        emit m_view.link_unhovered();
-    }
-
-    virtual void page_did_invalidate(Gfx::IntRect const&) override
-    {
-        m_view.viewport()->update();
-    }
-
-    virtual void page_did_change_favicon(Gfx::Bitmap const& bitmap) override
-    {
-        auto qimage = QImage(bitmap.scanline_u8(0), bitmap.width(), bitmap.height(), QImage::Format_ARGB32);
-        if (qimage.isNull())
-            return;
-        auto qpixmap = QPixmap::fromImage(qimage);
-        if (qpixmap.isNull())
-            return;
-        emit m_view.favicon_changed(QIcon(qpixmap));
-    }
-
-    virtual void page_did_layout() override
-    {
-        auto* layout_root = this->layout_root();
-        VERIFY(layout_root);
-        Gfx::IntSize content_size;
-        if (layout_root->paint_box()->has_overflow())
-            content_size = enclosing_int_rect(layout_root->paint_box()->scrollable_overflow_rect().value()).size();
-        else
-            content_size = enclosing_int_rect(layout_root->paint_box()->absolute_rect()).size();
-
-        m_view.verticalScrollBar()->setMaximum(content_size.height() - m_viewport_rect.height());
-        m_view.verticalScrollBar()->setPageStep(m_viewport_rect.height());
-        m_view.horizontalScrollBar()->setMaximum(content_size.width() - m_viewport_rect.width());
-        m_view.horizontalScrollBar()->setPageStep(m_viewport_rect.width());
-    }
-
-    virtual void page_did_request_scroll_into_view(Gfx::IntRect const& rect) override
-    {
-        if (m_viewport_rect.contains(rect))
-            return;
-
-        if (rect.top() < m_viewport_rect.top()) {
-            m_view.verticalScrollBar()->setValue(rect.top());
-        } else if (rect.top() > m_viewport_rect.top() && rect.bottom() > m_viewport_rect.bottom()) {
-            m_view.verticalScrollBar()->setValue(rect.bottom() - m_viewport_rect.height() + 1);
-        }
-    }
-
-    virtual void page_did_request_alert(String const& message) override
-    {
-        QMessageBox::warning(&m_view, "Ladybird", qstring_from_akstring(message));
-    }
-
-    virtual bool page_did_request_confirm(String const& message) override
-    {
-        auto result = QMessageBox::question(&m_view, "Ladybird", qstring_from_akstring(message),
-            QMessageBox::StandardButton::Ok | QMessageBox::StandardButton::Cancel);
-
-        return result == QMessageBox::StandardButton::Ok;
-    }
-
-    virtual String page_did_request_prompt(String const&, String const&) override
-    {
-        return String::empty();
-    }
-
-    virtual String page_did_request_cookie(AK::URL const& url, Web::Cookie::Source source) override
-    {
-        return m_cookie_jar.get_cookie(url, source);
-    }
-
-    virtual void page_did_set_cookie(AK::URL const& url, Web::Cookie::ParsedCookie const& cookie, Web::Cookie::Source source) override
-    {
-        m_cookie_jar.set_cookie(url, cookie, source);
-    }
-
-    void dump_cookies() const
-    {
-        m_cookie_jar.dump_cookies();
-    }
-
-    void request_file(NonnullRefPtr<Web::FileRequest>& request) override
-    {
-        auto const file = Core::System::open(request->path(), O_RDONLY);
-        request->on_file_request_finish(file);
-    }
-
-    void set_should_show_line_box_borders(bool state) { m_should_show_line_box_borders = state; }
-
-    HeadlessBrowserPageClient(WebView& view)
-        : m_view(view)
-        , m_page(make<Web::Page>(*this))
-    {
-    }
-
-    WebView& m_view;
-    NonnullOwnPtr<Web::Page> m_page;
-    Browser::CookieJar m_cookie_jar;
-
-    OwnPtr<Ladybird::ConsoleClient> m_console_client;
-    WeakPtr<JS::Realm> m_realm;
-    RefPtr<Gfx::PaletteImpl> m_palette_impl;
-    Gfx::IntRect m_viewport_rect { 0, 0, 800, 600 };
-    Web::CSS::PreferredColorScheme m_preferred_color_scheme { Web::CSS::PreferredColorScheme::Auto };
-    bool m_should_show_line_box_borders { false };
-};
 
 WebView::WebView()
 {
     setMouseTracking(true);
 
-    m_page_client = HeadlessBrowserPageClient::create(*this);
+    m_page_client = Ladybird::PageClientLadybird::create(*this);
 
     m_page_client->setup_palette(Gfx::load_system_theme(String::formatted("{}/res/themes/Default.ini", s_serenity_resource_root)));
 
@@ -818,7 +533,6 @@ void initialize_web_engine()
 
     Web::ResourceLoader::initialize(RequestManagerQt::create());
     Web::WebSockets::WebSocketClientManager::initialize(HeadlessWebSocketClientManager::create());
-
 
     Web::FrameLoader::set_default_favicon_path(String::formatted("{}/res/icons/16x16/app-browser.png", s_serenity_resource_root));
